==> notebooks/2_Analyse_des_corrélations.ipynb <==
# <span style="color:green">2. Analyse des corrélations</span>

L’objectif de cette étape est d’analyser les relations entre les variables après discrétisation afin de palier au problème de **multi-colinéarité** et préparer correctement les données pour les étapes suivantes du projet.

Cette étape utilise **exclusivement** le dataset issu de l’étape précédente :
**credit_risk_dataset_discretized_final.csv**

Il contient :
- Les **11 variables discrétisées** (ChiMerge, Monotone, KMeans)
- La variable cible : `loan_status`

 Pour vérifier la corrélation entre les variables discrétisées, et entre les variables explicatives et la variable à expliquer, on effectue le **test d’indépendance du X²**.
Le test d’indépendance du X² consiste à déterminer si la valeur observée d’une variable dépend de la valeur observée d’une autre variable. On visualise aussi avec la matrice de corrélation de Cramer.

#### <span style="color:orange">Librairies</span>

import sys
import pandas as pd
import plotly.express as px

sys.path.append("../src")
from utils import chi2_independence_tests, cramers_v_matrix

#### <span style="color:orange">Chargement du dataset</span> 

# 1. Chargement et Préparation
df = pd.read_csv('../datasets/processed/credit_risk_dataset_discretized_final.csv')
df.info()

# Séparation Features / Target
X = df.drop(columns=['loan_status'])
y = df['loan_status']

#### <span style="color:orange">Test d'indépendance de Chi-2</span>

Le test d’indépendance du χ² sert à déterminer si deux variables qualitatives (ou discrétisées) sont statistiquement indépendantes ou associées.

Hypothèses statistiques

Pour deux variables X1 et X2 :
- Hypothèse nulle : X1 et X2 sont indépendantes
- Hypothèse alternative X1 et X2 sont dépendantes 

test_chi2 = chi2_independence_tests(df.drop(columns=['loan_status', 'loan_grade']), alpha=0.05)
display(test_chi2.head())

Ces couples de variables présentent une dépendance statistiquement significative au seuil de 5%. 

test_chi2.to_excel("../result/test_chi2.xlsx", index=False)

#### <span style="color:orange">Matrice de corrélation avec V de Cramer</span>

Cramér's V est une mesure de l’association entre deux variables qualitatives (catégorielles). C'est l'équivalent du coefficient de corrélation de Pearson mais pour des catégories. Il utilise la statistique du test du chi² pour l’indépendance de deux variables. 
- 0≤V≤1 : PLus V tend vers 0 et moins les variables catégorielles sont liées



V_matrix = cramers_v_matrix(df).round(2)

# Visualisation avec Plotly
fig = px.imshow(
    V_matrix,
    text_auto=True,
    color_continuous_scale='Viridis',
    title="Matrice de corrélation Cramer's V",
    labels=dict(x="Variables", y="Variables", color="Cramér's V")
)
fig.update_layout(
    xaxis_side="top",
    width=800,
    height=800
)
fig.show()

Ces observations sont bien en adéquation avec les tets d'indépendance de Chi-2 et sont à prendre en compte pour la suite de l'analyse.
Les variables `person_income`,`loan_grade`,`cb_person_default_on_file` et `person_home_ownership` figurent parmi les plus correlées à notre target `loan_status`.

#### <span style="color:orange">Prochaine étape: La sélection de variables</span>
